Backend/requirements.txt
Django>=4.2
djangorestframework
PyMySQL
python-dotenv
django-cors-headers
